Hooked gameState to probability
Created the handleGameState function using a useCallback hook, so that the probabilityOfOutcome function would be called when the gameState changes (either the numberOfBalloons, or the balloonsPopped variables) and the result of the probbilityOfOutcome variable is displayed on the balloonPage. Next step is to make the probabilityOfOutcome function actually calculate the probability of the current state.