Implemented probabilityOfOutcome functionality
ProbabilityOfOutcome now determines the probability, given the number of balloons total, and the number of balloons that have been popped. We now display that probability, updating it with each change in the game state. Need to remove console.log statements, and decide if getting the 'tails' outcome should reset the entire game, or result in some other outcome (like a 'game over' message)