Slight optimization
Before, every time the Flip button was clicked, the functions being used to initialize the color, animation duration, and pop duration, of each balloon was rerunning. Replacing the randomX() with () => {return randomX()} makes this no longer happen. Now they only run on initialization, and when the useEffect calls them. (randomColor had to be put in a useCallback because it cares about the props.idNum value.)